# shopflow-monolith/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: ShopFlow API
  description: Event-driven e-commerce REST API — auth, products, inventory, and orders
  version: 1.0.0
  contact:
    name: Chandana Ganasala
    url: https://github.com/GanasalaChandana/Event-Driven-E-Commerce
servers:
  - url: https://event-driven-e-commerce.onrender.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItemRequest:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId: { type: string, format: uuid }
        productName: { type: string, example: Wireless Headphones }
        quantity: { type: integer, minimum: 1, example: 2 }
        unitPrice: { type: number, example: 99.99 }
    OrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemResponse:
      type: object
      properties:
        productId: { type: string, format: uuid }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        subtotal: { type: number }
    OrderResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        userEmail: { type: string }
        status: { type: string, enum: [PENDING, CONFIRMED, CANCELLED] }
        totalAmount: { type: number }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PageOfOrders:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order placed (status PENDING → CONFIRMED in ~10s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    get:
      tags: [Orders]
      summary: Get my orders (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfOrders'
  /api/v1/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel a PENDING order (only the order owner can cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found or not yours
        '409':
          description: Order is not in PENDING status
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get a specific order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not found